requirements.txt
pdf2image
pillow
google-generativeai
chromadb
openai
pandas
python-dotenv
retry
tiktoken
sentence_transformers
langchain